docs: Add Dialog README
